# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API HUB-UMKM Documentation",
        "description": "Documentation for HUB PHB UMKM",
        "contact": {
            "email": "jamal.apriadi@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get list of orders for the authenticated user",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "List of orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "total": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 250000
                                            },
                                            "shipping_address": {
                                                "type": "string",
                                                "example": "Jl. Merdeka 123"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-07-07T10:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "External Integration"
                ],
                "summary": "Terima data order dari sistem eksternal setelah pembayaran selesai",
                "operationId": "ab0100af245273306a225b01eaedfe55",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_number": {
                                        "type": "string",
                                        "example": "INV-64EE1F10A8E35"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "example": 250000
                                    },
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka 123"
                                    },
                                    "shipping_city": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "shipping_province": {
                                        "type": "string",
                                        "example": "DKI Jakarta"
                                    },
                                    "shipping_postal_code": {
                                        "type": "string",
                                        "example": "10110"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Bungkus rapi"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 125000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order berhasil diterima dari sistem eksternal"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order detail by ID",
                "operationId": "9fb6e018497aaebb6875182026ab3b9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 350000
                                        },
                                        "shipping_address": {
                                            "type": "string",
                                            "example": "Jl. Merdeka No. 123"
                                        },
                                        "order_items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product_name": {
                                                        "type": "string",
                                                        "example": "Laptop"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 15000000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/midtrans/callback": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Handle Midtrans payment notification callback",
                "operationId": "81752496c5efb10130a6c7dfd03b0215",
                "requestBody": {
                    "description": "Midtrans Notification Payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "transaction_status": {
                                        "type": "string",
                                        "example": "settlement"
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": "INV-64EE1F10A8E35"
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "example": "bank_transfer"
                                    },
                                    "fraud_status": {
                                        "type": "string",
                                        "example": "accept"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Callback handled successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
        "/api/product-categories/{slug}": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Get a product category by Slug",
                "operationId": "70c55c05abef01caeef2b3afaa20e6be",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/api/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by Slug",
                "operationId": "5bbb6d725346d38781b256ed5c715d62",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}